@@ -2,8 +2,8 @@ module ClientSideValidations::ActionView::Helpers
   module FormHelper
     class Error < StandardError; end
 
-    def form_for(record, *args, &proc)
-      options = args.extract_options!
+    def form_for(record, options = {}, &block)
+      raise ArgumentError, "Missing block" unless block_given?
       if options[:validate]
 
         # Always turn off HTML5 Validations
@@ -12,9 +12,10 @@ module ClientSideValidations::ActionView::Helpers
 
         case record
         when String, Symbol
-          raise ClientSideValidations::ActionView::Helpers::FormHelper::Error, 'Using form_for(:name, @resource) is not supported with ClientSideValidations. Please use form_for(@resource, :as => :name) instead.'
+          raise ClientSideValidations::ActionView::Helpers::FormHelper::Error, 'Using form_for(:name, @resource) is not supported with ClientSideValidations. Please use form_for(@resource, as: :name) instead.'
         else
           object = record.is_a?(Array) ? record.last : record
+          object_name = options[:as] || model_name_from_record_or_class(object).param_key
         end
       end
 
@@ -22,9 +23,12 @@ module ClientSideValidations::ActionView::Helpers
 
       # Order matters here. Rails mutates the options object
       html_id = options[:html][:id] if options[:html]
-      form   = super(record, *(args << options), &proc)
+      form = super(record, options, &block)
       options[:id] = html_id if html_id
-      script = client_side_form_settings(object, options)
+
+      build_bound_validators options
+      builder = instantiate_builder(object_name, object, options) if object_name and object
+      script = client_side_form_settings(object, options, builder)
 
       # Because of the load order requirement above this sub is necessary
       # Would be nice to not do this
@@ -44,13 +48,20 @@ module ClientSideValidations::ActionView::Helpers
       end
     end
 
-    def apply_form_for_options!(object_or_array, options)
+    def apply_form_for_options!(record, object, options)
       super
       options[:html][:validate] = true if options[:validate]
     end
 
     def fields_for(record_or_name_or_array, record_object = nil, options = {}, &block)
       output = super
+      build_bound_validators options
+      output
+    end
+
+    private
+
+    def build_bound_validators(options)
       if @validators
         options[:validators].each do |key, value|
           if @validators.key?(key)
@@ -60,10 +71,7 @@ module ClientSideValidations::ActionView::Helpers
           end
         end
       end
-      output
     end
-
-    private
 
     def insert_validators_into_script(script)
       # There is probably a more performant way of doing this
@@ -98,36 +106,30 @@ module ClientSideValidations::ActionView::Helpers
       end
     end
 
-    def client_side_form_settings(object, options)
+    def client_side_form_settings(object, options, builder)
       if options[:validate]
-        builder = options[:parent_builder]
-
         if options[:id]
           var_name = options[:id]
         else
-          if Rails.version >= '3.2.0'
-            var_name = if object.respond_to?(:persisted?) && object.persisted?
-              options[:as] ? "edit_#{options[:as]}" : [options[:namespace], dom_id(object, :edit)].compact.join("_")
-            else
-              options[:as] ? "new_#{options[:as]}" : [options[:namespace], dom_id(object)].compact.join("_")
-            end
+          var_name = if object.respond_to?(:persisted?) && object.persisted?
+            options[:as] ? "edit_#{options[:as]}" : [options[:namespace], dom_id(object, :edit)].compact.join("_")
           else
-            # This is to maintain backward compatibility with Rails 3.1
-            # see: https://github.com/rails/rails/commit/e29773f885fd500189ffd964550ae20061d745ba#commitcomment-948052
-            var_name = if object.respond_to?(:persisted?) && object.persisted?
-              options[:as] ? "#{options[:as]}_edit" : dom_id(object, :edit)
-            else
-              options[:as] ? "#{options[:as]}_new" : dom_id(object)
-            end
+            options[:as] ? "new_#{options[:as]}" : [options[:namespace], dom_id(object)].compact.join("_")
           end
         end
 
+        if ClientSideValidations::Config.number_format_with_locale and defined?(I18n)
+          number_format = I18n.t('number.format').slice(:separator, :delimiter)
+        else
+          number_format = { separator:'.', delimiter:',' }
+        end
+        patterns = { numericality:"/^(-|\\+)?(?:\\d+|\\d{1,3}(?:\\#{number_format[:delimiter]}\\d{3})+)(?:\\#{number_format[:separator]}\\d*)?$/" }
+
+
         content_tag(:script) do
-          "//<![CDATA[\nif(window.ClientSideValidations==undefined)window.ClientSideValidations={};if(window.ClientSideValidations.remote_validators_prefix==undefined)window.ClientSideValidations.remote_validators_prefix='#{(root_path||"").sub(/(\/)+$/,'')}';if(window.ClientSideValidations.forms==undefined)window.ClientSideValidations.forms={};window.ClientSideValidations.forms['#{var_name}'] = #{builder.client_side_form_settings(options, self).merge(:validators => 'validator_hash').to_json};\n//]]>".html_safe
+          "//<![CDATA[\nif(window.ClientSideValidations===undefined)window.ClientSideValidations={};window.ClientSideValidations.disabled_validators=#{ClientSideValidations::Config.disabled_validators.to_json};window.ClientSideValidations.number_format=#{number_format.to_json};if(window.ClientSideValidations.patterns===undefined)window.ClientSideValidations.patterns = {};window.ClientSideValidations.patterns.numericality=#{patterns[:numericality]};#{"if(window.ClientSideValidations.remote_validators_prefix===undefined)window.ClientSideValidations.remote_validators_prefix='#{(ClientSideValidations::Config.root_path).sub(/\/+\Z/,'')}';" if ClientSideValidations::Config.root_path.present? }if(window.ClientSideValidations.forms===undefined)window.ClientSideValidations.forms={};window.ClientSideValidations.forms['#{var_name}'] = #{builder.client_side_form_settings(options, self).merge(validators: 'validator_hash').to_json};\n//]]>".html_safe
         end
       end
     end
-
   end
 end
-
