@@ -14,7 +14,7 @@ module ClientSideValidations::ActiveModel
     private
 
     def build_client_side_hash(model, attribute, options)
-      { :message => model.errors.generate_message(attribute, message_type, options) }.merge(options.except(*::ActiveModel::Errors::CALLBACKS_OPTIONS - [:allow_blank, :if, :unless]))
+      { message: model.errors.generate_message(attribute, message_type, options) }.merge(options.except(*::ActiveModel::Errors::CALLBACKS_OPTIONS - [:allow_blank, :if, :unless]))
     end
 
     def message_type
@@ -70,7 +70,9 @@ module ClientSideValidations::ActiveModel
         result = result && validator.kind != :block
 
         if validator.options[:if] || validator.options[:unless]
-          if result = can_force_validator?(attr, validator, force)
+          if validator.options[:if] && validator.options[:if] =~ /changed\?/
+            result = true
+          else result = can_force_validator?(attr, validator, force)
             if validator.options[:if]
               result = result && run_conditional(validator.options[:if])
             end
@@ -84,12 +86,11 @@ module ClientSideValidations::ActiveModel
       result
     end
 
-    def run_conditional(method_name_or_proc)
-      case method_name_or_proc
-      when Proc
-        method_name_or_proc.call(self)
+    def run_conditional(method_name_value_or_proc)
+      if method_name_value_or_proc.respond_to?(:call)
+        method_name_value_or_proc.call(self)
       else
-        self.send(method_name_or_proc)
+        self.send(method_name_value_or_proc)
       end
     end
 
@@ -125,11 +126,7 @@ module ClientSideValidations::ActiveModel
           false
         end
       else
-        if (validator.options[:if] || validator.options[:unless]) =~ /changed\?/
-          true
-        else
-          false
-        end
+        false
       end
     end
   end
@@ -138,9 +135,8 @@ end
 ActiveModel::Validator.send(:include, ClientSideValidations::ActiveModel::Validator)
 ActiveModel::Validations.send(:include, ClientSideValidations::ActiveModel::Validations)
 
-%w{acceptance exclusion inclusion length format numericality presence}.each do |validator|
+%w{absence acceptance exclusion inclusion length format numericality presence}.each do |validator|
   require "client_side_validations/active_model/#{validator}"
   validator.capitalize!
   eval "ActiveModel::Validations::#{validator}Validator.send(:include, ClientSideValidations::ActiveModel::#{validator})"
 end
-
