@@ -3,7 +3,7 @@ module ClientSideValidations::ActiveModel
 
     def client_side_hash(model, attribute, force = nil)
       options = self.options.dup
-      hash    = { :messages => {} }
+      hash    = { messages: {} }
       hash[:js_tokenizer] = options[:js_tokenizer] if options[:js_tokenizer]
       hash[:allow_blank]  = true if options[:allow_blank]
 
@@ -11,7 +11,7 @@ module ClientSideValidations::ActiveModel
         if count = options[option]
           options[:message] = options[message_type] if options[message_type].present?
           options.delete(:message) if options[:message].nil?
-          hash[:messages][option] = model.errors.generate_message(attribute, message_type, options.merge(:count => count))
+          hash[:messages][option] = model.errors.generate_message(attribute, message_type, options.merge(count: count))
           hash[option] = count
         end
       end
@@ -23,4 +23,3 @@ module ClientSideValidations::ActiveModel
 
   end
 end
-
