@@ -19,57 +19,51 @@ module ActionViewTestSetup
   end
 
   Routes = ActionDispatch::Routing::RouteSet.new
+  include Routes.url_helpers
+  def _routes
+    Routes
+  end
+
   Routes.draw do
     resources :posts do
       resources :comments
     end
 
-    root :to => 'main#index'
+    root to: 'main#index'
+  end
+
+  def default_url_options
+    { only_path: true }
   end
 
   def url_for(object)
     @url_for_options = object
     if object.is_a?(Hash) && object[:use_route].blank? && object[:controller].blank?
-      object.merge!(:controller => "main", :action => "index")
+      object.merge!(controller: "main", action: "index")
     end
     super
   end
 
-  def _routes
-    Routes
-  end
-
-  # Rails 3.2.0 dropped size from the form elements
-  def legacy_size
-    if Rails.version < '3.2.0'
-      'size="30" '
-    end
-  end
-
   def hidden_input_for_select(name)
-    if Rails.version >= '3.2.0'
-      %{<input name="#{name}" type="hidden" value="" />}
-    end
+    %{<input name="#{name}" type="hidden" value="" />}
   end
-
-  include Routes.url_helpers
 
   def setup
     super
 
     # Create 'label' locale for testing I18n label helpers
     I18n.backend.store_translations 'label', {
-      :activemodel => {
-        :attributes => {
-          :post => {
-            :cost => 'Total cost'
+      activemodel: {
+        attributes: {
+          post: {
+            cost: 'Total cost'
           }
         }
       },
-      :helpers => {
-        :label => {
-          :post => {
-            :body => 'Write entire text here'
+      helpers: {
+        label: {
+          post: {
+            body: 'Write entire text here'
           }
         }
       }
@@ -77,13 +71,13 @@ module ActionViewTestSetup
 
     # Create "submit" locale for testing I18n submit helpers
     I18n.backend.store_translations 'submit', {
-      :helpers => {
-        :submit => {
-          :create => 'Create %{model}',
-          :update => 'Confirm %{model} changes',
-          :submit => 'Save changes',
-          :another_post => {
-            :update => 'Update your %{model}'
+      helpers: {
+        submit: {
+          create: 'Create %{model}',
+          update: 'Confirm %{model} changes',
+          submit: 'Save changes',
+          another_post: {
+            update: 'Update your %{model}'
           }
         }
       }
@@ -100,34 +94,57 @@ module ActionViewTestSetup
   end
 
   def snowman(method = nil)
-    txt =  %{<div style="margin:0;padding:0;display:inline">}
+    txt =
+      if Rails.version.starts_with?('4.0')
+        %{<div style="margin:0;padding:0;display:inline">}
+      elsif Rails.version.starts_with?('4.1')
+        %{<div style="display:none">}
+      else
+        ''
+      end
     txt << %{<input name="utf8" type="hidden" value="&#x2713;" />}
-    txt << %{<input name="_method" type="hidden" value="#{method}" />} if method
-    txt << %{</div>}
+    txt << %{<input type="hidden" name="_method" value="#{method}" />} if method
+    txt << %{</div>} unless Rails.version.starts_with?('4.2')
+    txt
   end
 
-  def form_text(action = "http://www.example.com", id = nil, html_class = nil, remote = nil, validators = nil, file = nil)
-    txt =  %{<form accept-charset="UTF-8" action="#{action}"}
-    txt << %{ data-remote="true"} if remote
-    txt << %{ class="#{html_class}"} if html_class
+  def form_text(action = "http://www.example.com", id = nil, html_class = nil, remote = nil, validators = nil, file = nil, custom_id = false)
+    txt =  %{<form action="#{action}" accept-charset="UTF-8" method="post"}
     txt << %{ data-validate="true"} if validators
-    txt << %{ enctype="multipart/form-data"} if file
-    txt << %{ id="#{id}"} if id
-    txt << %{ method="post"}
+    txt << %{ id="#{id}"} if id && custom_id
     txt << %{ novalidate="novalidate"} if validators
+    txt << %{ class="#{html_class}"} if html_class
+    txt << %{ id="#{id}"} if id && !custom_id
+    txt << %{ enctype="multipart/form-data"} if file
     txt << %{>}
+  end
+
+  def form_field(tag, id = nil, name = nil, type = nil, value = nil, multiple = false, tag_content = nil, custom_name = nil)
+    txt =  %{<#{tag}}
+    txt << %{ name="#{custom_name}"} if custom_name
+    txt << %{ type="#{type}"} if type
+    txt << %{ value="#{value}"} if value
+    txt << %{ multiple="multiple"} if multiple
+    txt << %{ name="#{name}"} if name
+    txt << %{ id="#{id}"} if id
+    txt <<
+       if %w(select textarea).include?(tag)
+         %{>#{tag_content}</#{tag}>}
+       else
+         %{ />}
+       end
   end
 
   def whole_form(action = "http://www.example.com", id = nil, html_class = nil, options = nil)
     contents = block_given? ? yield : ""
 
     if options.is_a?(Hash)
-      method, remote, validators, file = options.values_at(:method, :remote, :validators, :file)
+      method, remote, validators, file, custom_id, no_validate = options.values_at(:method, :remote, :validators, :file, :custom_id, :no_validate)
     else
       method = options
     end
 
-    html = form_text(action, id, html_class, remote, validators, file) + snowman(method) + (contents || "") + "</form>"
+    html = form_text(action, id, html_class, remote, (validators || no_validate), file, custom_id) + snowman(method) + (contents || "") + "</form>"
 
     if options.is_a?(Hash) && options[:validators]
       build_script_tag(html, id, options[:validators])
@@ -137,7 +154,9 @@ module ActionViewTestSetup
   end
 
   def build_script_tag(html, id, validators)
-    (html || "") + %Q{<script>//<![CDATA[\nif(window.ClientSideValidations==undefined)window.ClientSideValidations={};if(window.ClientSideValidations.remote_validators_prefix==undefined)window.ClientSideValidations.remote_validators_prefix='';if(window.ClientSideValidations.forms==undefined)window.ClientSideValidations.forms={};window.ClientSideValidations.forms['#{id}'] = #{client_side_form_settings_helper.merge(:validators => validators).to_json};\n//]]></script>}
+    number_format = {separator: '.', delimiter: ','}
+    patterns = {numericality:"/^(-|\\+)?(?:\\d+|\\d{1,3}(?:\\#{number_format[:delimiter]}\\d{3})+)(?:\\#{number_format[:separator]}\\d*)?$/"}
+    (html || '') + %Q{<script>//<![CDATA[\nif(window.ClientSideValidations===undefined)window.ClientSideValidations={};window.ClientSideValidations.disabled_validators=#{ClientSideValidations::Config.disabled_validators.to_json};window.ClientSideValidations.number_format=#{number_format.to_json};if(window.ClientSideValidations.patterns===undefined)window.ClientSideValidations.patterns = {};window.ClientSideValidations.patterns.numericality=#{patterns[:numericality]};if(window.ClientSideValidations.forms===undefined)window.ClientSideValidations.forms={};window.ClientSideValidations.forms['#{id}'] = #{client_side_form_settings_helper.merge(validators: validators).to_json};\n//]]></script>}
   end
 
   protected
@@ -176,6 +195,4 @@ module ActionViewTestSetup
     def protect_against_forgery?
       false
     end
-
 end
-
