@@ -9,9 +9,9 @@ class ClientSideValidations::LegacyActionViewHelpersTest < ActionView::TestCase
     end
 
     expected = whole_form('/posts', 'new_post', 'new_post') do
-      %{<input id="post_cost" name="post[cost]" size="30" type="text" />}
+      form_field('input', 'post_cost', 'post[cost]', 'text')
     end
-    assert_equal expected, output_buffer
+    assert_dom_equal expected, output_buffer
   end
 
   def test_password_field
@@ -20,44 +20,31 @@ class ClientSideValidations::LegacyActionViewHelpersTest < ActionView::TestCase
     end
 
     expected = whole_form('/posts', 'new_post', 'new_post') do
-      %{<input id="post_cost" name="post[cost]" size="30" type="password" />}
+      form_field('input', 'post_cost', 'post[cost]', 'password')
     end
-    assert_equal expected, output_buffer
+    assert_dom_equal expected, output_buffer
   end
 
   def test_file_field
-    form_for(@post, :html => {:multipart => true}) do |f|
+    form_for(@post, html: {multipart: true}) do |f|
       concat f.file_field(:cost)
     end
 
-    expected = whole_form('/posts', 'new_post', 'new_post', :file => true) do
-      %{<input id="post_cost" name="post[cost]" type="file" />}
+    expected = whole_form('/posts', 'new_post', 'new_post', file: true) do
+      form_field('input', 'post_cost', 'post[cost]', 'file')
     end
-    assert_equal expected, output_buffer
+    assert_dom_equal expected, output_buffer
   end
 
-  if Rails.version >= '3.2.0'
-    def test_text_area
-      form_for(@post) do |f|
-        concat f.text_area(:cost)
-      end
-
-      expected = whole_form('/posts', 'new_post', 'new_post') do
-        %{<textarea cols="40" id="post_cost" name="post[cost]" rows="20">\n</textarea>}
-      end
-      assert_equal expected, output_buffer
+  def test_text_area
+    form_for(@post) do |f|
+      concat f.text_area(:cost)
     end
-  elsif Rails.version >= '3.1.0'
-    def test_text_area
-      form_for(@post) do |f|
-        concat f.text_area(:cost)
-      end
 
-      expected = whole_form('/posts', 'new_post', 'new_post') do
-        %{<textarea cols="40" id="post_cost" name="post[cost]" rows="20"></textarea>}
-      end
-      assert_equal expected, output_buffer
+    expected = whole_form('/posts', 'new_post', 'new_post') do
+      form_field('textarea', 'post_cost', 'post[cost]', nil, nil, nil, "\n")
     end
+    assert_dom_equal expected, output_buffer
   end
 
   def test_search_field
@@ -66,9 +53,9 @@ class ClientSideValidations::LegacyActionViewHelpersTest < ActionView::TestCase
     end
 
     expected = whole_form('/posts', 'new_post', 'new_post') do
-      %{<input id="post_cost" name="post[cost]" size="30" type="search" />}
+      form_field('input', 'post_cost', 'post[cost]', 'search')
     end
-    assert_equal expected, output_buffer
+    assert_dom_equal expected, output_buffer
   end
 
   def test_telephone_field
@@ -77,9 +64,9 @@ class ClientSideValidations::LegacyActionViewHelpersTest < ActionView::TestCase
     end
 
     expected = whole_form('/posts', 'new_post', 'new_post') do
-      %{<input id="post_cost" name="post[cost]" size="30" type="tel" />}
+      form_field('input', 'post_cost', 'post[cost]', 'tel')
     end
-    assert_equal expected, output_buffer
+    assert_dom_equal expected, output_buffer
   end
 
   def test_phone_field
@@ -88,9 +75,9 @@ class ClientSideValidations::LegacyActionViewHelpersTest < ActionView::TestCase
     end
 
     expected = whole_form('/posts', 'new_post', 'new_post') do
-      %{<input id="post_cost" name="post[cost]" size="30" type="tel" />}
+      form_field('input', 'post_cost', 'post[cost]', 'tel')
     end
-    assert_equal expected, output_buffer
+    assert_dom_equal expected, output_buffer
   end
 
   def test_url_field
@@ -99,9 +86,9 @@ class ClientSideValidations::LegacyActionViewHelpersTest < ActionView::TestCase
     end
 
     expected = whole_form('/posts', 'new_post', 'new_post') do
-      %{<input id="post_cost" name="post[cost]" size="30" type="url" />}
+      form_field('input', 'post_cost', 'post[cost]', 'url')
     end
-    assert_equal expected, output_buffer
+    assert_dom_equal expected, output_buffer
   end
 
   def test_email_field
@@ -110,9 +97,9 @@ class ClientSideValidations::LegacyActionViewHelpersTest < ActionView::TestCase
     end
 
     expected = whole_form('/posts', 'new_post', 'new_post') do
-      %{<input id="post_cost" name="post[cost]" size="30" type="email" />}
+      form_field('input', 'post_cost', 'post[cost]', 'email')
     end
-    assert_equal expected, output_buffer
+    assert_dom_equal expected, output_buffer
   end
 
   def test_number_field
@@ -121,9 +108,9 @@ class ClientSideValidations::LegacyActionViewHelpersTest < ActionView::TestCase
     end
 
     expected = whole_form('/posts', 'new_post', 'new_post') do
-      %{<input id="post_cost" name="post[cost]" #{legacy_size}type="number" />}
+      form_field('input', 'post_cost', 'post[cost]', 'number')
     end
-    assert_equal expected, output_buffer
+    assert_dom_equal expected, output_buffer
   end
 
   def test_range_field
@@ -132,9 +119,9 @@ class ClientSideValidations::LegacyActionViewHelpersTest < ActionView::TestCase
     end
 
     expected = whole_form('/posts', 'new_post', 'new_post') do
-      %{<input id="post_cost" name="post[cost]" #{legacy_size}type="range" />}
+      form_field('input', 'post_cost', 'post[cost]', 'range')
     end
-    assert_equal expected, output_buffer
+    assert_dom_equal expected, output_buffer
   end
 
   def test_check_box
@@ -144,9 +131,9 @@ class ClientSideValidations::LegacyActionViewHelpersTest < ActionView::TestCase
 
     expected = whole_form('/posts', 'new_post', 'new_post') do
       %{<input name="post[cost]" type="hidden" value="0" />} +
-      %{<input id="post_cost" name="post[cost]" type="checkbox" value="1" />}
+      form_field('input', 'post_cost', 'post[cost]', 'checkbox', '1')
     end
-    assert_equal expected, output_buffer
+    assert_dom_equal expected, output_buffer
   end
 
   def test_radio_button
@@ -155,9 +142,9 @@ class ClientSideValidations::LegacyActionViewHelpersTest < ActionView::TestCase
     end
 
     expected = whole_form('/posts', 'new_post', 'new_post') do
-      %{<input id="post_cost_10" name="post[cost]" type="radio" value="10" />}
+      form_field('input', 'post_cost_10', 'post[cost]', 'radio', '10')
     end
-    assert_equal expected, output_buffer
+    assert_dom_equal expected, output_buffer
   end
 
   def test_fields_for
@@ -165,9 +152,9 @@ class ClientSideValidations::LegacyActionViewHelpersTest < ActionView::TestCase
       c.text_field(:title)
     end
 
-    expected = %{<input id="comment_title" name="comment[title]" size="30" type="text" />}
+    expected = form_field('input', 'comment_title', 'comment[title]', 'text')
 
-    assert_equal expected, result
+    assert_dom_equal expected, result
   end
 
   def test_select
@@ -176,20 +163,20 @@ class ClientSideValidations::LegacyActionViewHelpersTest < ActionView::TestCase
     end
 
     expected = whole_form('/posts', 'new_post', 'new_post') do
-      %{<select id="post_cost" name="post[cost]"></select>}
+      form_field('select', 'post_cost', 'post[cost]')
     end
-    assert_equal expected, output_buffer
+    assert_dom_equal expected, output_buffer
   end
 
   def test_select_multiple
     form_for(@post) do |f|
-      concat f.select(:cost, [], {}, :multiple => true)
+      concat f.select(:cost, [], {}, multiple: true)
     end
 
     expected = whole_form('/posts', 'new_post', 'new_post') do
-      %{#{hidden_input_for_select('post[cost][]')}<select id="post_cost" multiple="multiple" name="post[cost][]"></select>}
+      %{#{hidden_input_for_select('post[cost][]')}#{form_field('select', 'post_cost', 'post[cost][]', nil, nil, true)}}
     end
-    assert_equal expected, output_buffer
+    assert_dom_equal expected, output_buffer
   end
 
   def test_collection_select
@@ -198,9 +185,9 @@ class ClientSideValidations::LegacyActionViewHelpersTest < ActionView::TestCase
     end
 
     expected = whole_form('/posts', 'new_post', 'new_post') do
-      %{<select id="post_cost" name="post[cost]"></select>}
+      form_field('select', 'post_cost', 'post[cost]')
     end
-    assert_equal expected, output_buffer
+    assert_dom_equal expected, output_buffer
   end
 
   def test_grouped_collection_select
@@ -209,22 +196,22 @@ class ClientSideValidations::LegacyActionViewHelpersTest < ActionView::TestCase
     end
 
     expected = whole_form('/posts', 'new_post', 'new_post') do
-      %{<select id="post_cost" name="post[cost]"></select>}
+      form_field('select', 'post_cost', 'post[cost]')
     end
-    assert_equal expected, output_buffer
+    assert_dom_equal expected, output_buffer
   end
 
   def test_time_zone_select
     zones = mock('TimeZones')
     zones.stubs(:all).returns([])
     form_for(@post) do |f|
-      concat f.time_zone_select(:cost, nil, :model => zones)
+      concat f.time_zone_select(:cost, nil, model: zones)
     end
 
     expected = whole_form('/posts', 'new_post', 'new_post') do
-      %{<select id="post_cost" name="post[cost]"></select>}
+      form_field('select', 'post_cost', 'post[cost]')
     end
-    assert_equal expected, output_buffer
+    assert_dom_equal expected, output_buffer
   end
 
   def test_string_as_record
@@ -233,9 +220,9 @@ class ClientSideValidations::LegacyActionViewHelpersTest < ActionView::TestCase
     end
 
     expected = whole_form('/') do
-      %{<input id="post_cost" name="post[cost]" size="30" type="text" />}
+      form_field('input', 'post_cost', 'post[cost]', 'text')
     end
-    assert_equal expected, output_buffer
+    assert_dom_equal expected, output_buffer
   end
 
   def test_symbol_as_record
@@ -244,9 +231,8 @@ class ClientSideValidations::LegacyActionViewHelpersTest < ActionView::TestCase
     end
 
     expected = whole_form('/') do
-      %{<input id="post_cost" name="post[cost]" size="30" type="text" />}
+      form_field('input', 'post_cost', 'post[cost]', 'text')
     end
-    assert_equal expected, output_buffer
+    assert_dom_equal expected, output_buffer
   end
 end
-
