@@ -5,8 +5,8 @@ class Post
   include ActiveModel::Conversion
 
   attr_accessor :title, :author_name, :body, :secret, :written_on, :cost
-  validates :cost, :body, :presence => true
-  validates :body, :length => { :minimum => 200 }
+  validates :cost, :body, presence: true
+  validates :body, length: { minimum: 200 }
 
   def initialize(params={})
     params.each do |attr, value|
@@ -21,4 +21,3 @@ class Post
   attr_accessor :comments, :comment_ids
   def comments_attributes=(attributes); end
 end
-
