@@ -3,24 +3,11 @@
 require 'middleware/cases/helper'
 require 'active_record/cases/helper'
 
-class ClientSideValidationsActiveRecordMiddlewareTest < Test::Unit::TestCase
+class ClientSideValidationsActiveRecordMiddlewareTest < MiniTest::Test
   include Rack::Test::Methods
 
   def teardown
     User.delete_all
-  end
-
-  def with_kcode(kcode)
-    if RUBY_VERSION < '1.9'
-      orig_kcode, $KCODE = $KCODE, kcode
-      begin
-        yield
-      ensure
-        $KCODE = orig_kcode
-      end
-    else
-      yield
-    end
   end
 
   def app
@@ -29,7 +16,7 @@ class ClientSideValidationsActiveRecordMiddlewareTest < Test::Unit::TestCase
   end
 
   def test_uniqueness_when_resource_exists
-    User.create(:email => 'user@test.com')
+    User.create(email: 'user@test.com')
     get '/validators/uniqueness', { 'user[email]' => 'user@test.com', 'case_sensitive' => true }
 
     assert_equal 'false', last_response.body
@@ -37,7 +24,7 @@ class ClientSideValidationsActiveRecordMiddlewareTest < Test::Unit::TestCase
   end
 
   def test_uniqueness_when_resource_exists_and_param_order_is_backwards
-    User.create(:email => 'user@test.com')
+    User.create(email: 'user@test.com')
     get '/validators/uniqueness', { 'case_sensitive' => true, 'user[email]' => 'user@test.com' }
 
     assert_equal 'false', last_response.body
@@ -52,7 +39,7 @@ class ClientSideValidationsActiveRecordMiddlewareTest < Test::Unit::TestCase
   end
 
   def test_uniqueness_when_value_must_be_typecast
-    User.create(:active => false)
+    User.create(active: false)
     get '/validators/uniqueness', { 'user[active]' => 'false', 'case_sensitive' => true }
 
     assert_equal 'false', last_response.body
@@ -60,7 +47,7 @@ class ClientSideValidationsActiveRecordMiddlewareTest < Test::Unit::TestCase
   end
 
   def test_uniqueness_when_id_is_given
-    user = User.create(:email => 'user@test.com')
+    user = User.create(email: 'user@test.com')
     get '/validators/uniqueness', { 'user[email]' => 'user@test.com', 'id' => user.id, 'case_sensitive' => true }
 
     assert_equal 'true', last_response.body
@@ -68,7 +55,7 @@ class ClientSideValidationsActiveRecordMiddlewareTest < Test::Unit::TestCase
   end
 
   def test_mysql_adapter_uniqueness_when_id_is_given
-    user = User.create(:email => 'user@test.com')
+    user = User.create(email: 'user@test.com')
     User.connection.stubs('adapter_name').returns('Mysql2')
 
     sql_without_binary = "#{User.arel_table["email"].eq(user.email).to_sql} AND #{User.arel_table.primary_key.not_eq(user.id).to_sql}"
@@ -83,7 +70,7 @@ class ClientSideValidationsActiveRecordMiddlewareTest < Test::Unit::TestCase
   end
 
   def test_mysql_adapter_uniqueness_when_id_is_given_with_scope
-    user = User.create(:email => 'user@test.com', :name => 'Brian')
+    user = User.create(email: 'user@test.com', name: 'Brian')
     User.connection.stubs('adapter_name').returns('Mysql2')
 
     sql_without_binary = "#{User.arel_table["email"].eq(user.email).to_sql} AND #{User.arel_table.primary_key.not_eq(user.id).to_sql} AND #{User.arel_table["name"].eq(user.name).to_sql}"
@@ -98,7 +85,7 @@ class ClientSideValidationsActiveRecordMiddlewareTest < Test::Unit::TestCase
   end
 
   def test_uniqueness_when_scope_is_given
-    User.create(:email => 'user@test.com', :age => 25)
+    User.create(email: 'user@test.com', age: 25)
     get '/validators/uniqueness', { 'user[email]' => 'user@test.com', 'scope' => { 'age' => 30 }, 'case_sensitive' => true }
 
     assert_equal 'true', last_response.body
@@ -106,7 +93,7 @@ class ClientSideValidationsActiveRecordMiddlewareTest < Test::Unit::TestCase
   end
 
   def test_uniqueness_when_scope_is_given_and_value_must_be_typecast
-    User.create(:email => 'user@test.com', :active => true)
+    User.create(email: 'user@test.com', active: true)
     get '/validators/uniqueness', { 'user[email]' => 'user@test.com', 'scope' => { 'active' => 'true' }, 'case_sensitive' => true }
 
     assert_equal 'false', last_response.body
@@ -114,7 +101,7 @@ class ClientSideValidationsActiveRecordMiddlewareTest < Test::Unit::TestCase
   end
 
   def test_uniqueness_when_multiple_scopes_are_given
-    User.create(:email => 'user@test.com', :age => 30, :name => 'Brian')
+    User.create(email: 'user@test.com', age: 30, name: 'Brian')
     get '/validators/uniqueness', { 'user[email]' => 'user@test.com', 'scope' => { 'age' => 30, 'name' => 'Robert' }, 'case_sensitive' => true }
 
     assert_equal 'true', last_response.body
@@ -122,7 +109,7 @@ class ClientSideValidationsActiveRecordMiddlewareTest < Test::Unit::TestCase
   end
 
   def test_uniqueness_when_case_insensitive
-    User.create(:name => 'Brian')
+    User.create(name: 'Brian')
     get '/validators/uniqueness', { 'user[name]' => 'BRIAN', 'case_sensitive' => false }
 
     assert_equal 'false', last_response.body
@@ -130,7 +117,7 @@ class ClientSideValidationsActiveRecordMiddlewareTest < Test::Unit::TestCase
   end
 
   def test_uniqueness_when_attribute_passes_as_an_integer
-    User.create(:name => 123)
+    User.create(name: 123)
     get '/validators/uniqueness', { 'user[name]' => 123, 'case_sensitive' => true }
 
     assert_equal 'false', last_response.body
@@ -138,7 +125,7 @@ class ClientSideValidationsActiveRecordMiddlewareTest < Test::Unit::TestCase
   end
 
   def test_uniqueness_when_attribute_passes_as_an_integer
-    User.create(:name => 123)
+    User.create(name: 123)
     get '/validators/uniqueness', { 'user[name]' => 123, 'case_sensitive' => true }
 
     assert_equal 'false', last_response.body
@@ -147,14 +134,12 @@ class ClientSideValidationsActiveRecordMiddlewareTest < Test::Unit::TestCase
 
   def test_uniqueness_with_columns_which_are_sql_keywords
     Guid.validates_uniqueness_of :key
-    assert_nothing_raised do
-      get '/validators/uniqueness', { 'guid[key]' => 'test', 'case_sensitive' => true }
-    end
+    get '/validators/uniqueness', { 'guid[key]' => 'test', 'case_sensitive' => true }
   end
 
   def test_uniqueness_with_limit
     # User.title is limited to 5 characters
-    User.create(:title => "abcde")
+    User.create(title: "abcde")
     get '/validators/uniqueness', { 'user[title]' => 'abcdefgh', 'case_sensitive' => true }
 
     assert_equal 'false', last_response.body
@@ -162,14 +147,12 @@ class ClientSideValidationsActiveRecordMiddlewareTest < Test::Unit::TestCase
   end
 
   def test_uniqueness_with_limit_and_utf8
-    with_kcode('UTF8') do
-      # User.title is limited to 5 characters
-      User.create(:title => "一二三四五")
-      get '/validators/uniqueness', { 'user[title]' => '一二三四五六七八', 'case_sensitive' => true }
+    # User.title is limited to 5 characters
+    User.create(title: "一二三四五")
+    get '/validators/uniqueness', { 'user[title]' => '一二三四五六七八', 'case_sensitive' => true }
 
-      assert_equal 'false', last_response.body
+    assert_equal 'false', last_response.body
     assert last_response.ok?
-    end
   end
 
   def test_validate_straight_inheritance_uniqueness
@@ -177,7 +160,7 @@ class ClientSideValidationsActiveRecordMiddlewareTest < Test::Unit::TestCase
     assert_equal 'true', last_response.body
     assert last_response.not_found?
 
-    IneptWizard.create(:name => 'Rincewind')
+    IneptWizard.create(name: 'Rincewind')
     get '/validators/uniqueness', { 'inept_wizard[name]' => 'Rincewind', 'case_sensitive' => true }
     assert_equal 'false', last_response.body
     assert last_response.ok?
@@ -186,18 +169,26 @@ class ClientSideValidationsActiveRecordMiddlewareTest < Test::Unit::TestCase
     assert_equal 'false', last_response.body
     assert last_response.ok?
 
-    Conjurer.create(:name => 'The Amazing Bonko')
+    Conjurer.create(name: 'The Amazing Bonko')
     get '/validators/uniqueness', { 'thaumaturgist[name]' => 'The Amazing Bonko', 'case_sensitive' => true }
     assert_equal 'false', last_response.body
     assert last_response.ok?
   end
 
   def test_uniqueness_when_resource_is_a_nested_module
-    ActiveRecordTestModule::User2.create(:email => 'user@test.com')
+    ActiveRecordTestModule::User2.create(email: 'user@test.com')
     get '/validators/uniqueness', { 'active_record_test_module/user2[email]' => 'user@test.com', 'case_sensitive' => true }
 
     assert_equal 'false', last_response.body
     assert last_response.ok?
   end
-end
 
+  def test_uniqueness_when_resource_descends_from_an_abstract_base_class
+      Thing.create(name: 'name_to_be_duplicated')
+      get '/validators/uniqueness', { 'thing[name]' => 'name_to_be_duplicated', 'case_sensitive' => true }
+
+      assert_equal 'false', last_response.body
+      assert last_response.ok?
+  end
+
+end
