@@ -2,7 +2,7 @@ users_table = %{CREATE TABLE users (id INTEGER PRIMARY KEY, age INTEGER, name TE
 ActiveRecord::Base.connection.execute(users_table)
 
 class User < ActiveRecord::Base
-  validates :email, :title, :active, :name, :uniqueness => { :allow_nil => true }
+  validates :email, :title, :active, :name, uniqueness: { allow_nil: true }
 end
 
 class IneptWizard < User; end
@@ -11,4 +11,20 @@ class Thaumaturgist < Conjurer; end
 
 module ActiveRecordTestModule
   class User2 < User; end
+end
+
+class UserForm
+  include ActiveRecord::Validations
+
+  attr_accessor :name
+
+  validates_uniqueness_of :name, client_validations: { class: User }
+
+  def self.i18n_scope
+    :activerecord
+  end
+
+  def new_record?
+    true
+  end
 end
