@@ -1,3 +1,3 @@
 module ClientSideValidations
-  VERSION = '3.2.1'
+  VERSION = '4.2.0'.freeze
 end
