@@ -1,26 +1,28 @@
 require 'base_helper'
 require 'client_side_validations/core_ext'
 
-class CoreExtTest < Test::Unit::TestCase
-  def test_regexp_as_json
-    regexp = //
-    assert_equal regexp, regexp.as_json
-  end
-
+class CoreExtTest < MiniTest::Test
   def test_regexp_replace_A_and_Z
     test_regexp = /\A\Z/
-    expected_regexp = /^$/
+    expected_regexp = { source: '^$', options: '' }
     assert_equal expected_regexp, test_regexp.as_json
   end
 
   def test_regexp_replace_a_and_z
     test_regexp = /\A\z/
-    expected_regexp = /^$/
+    expected_regexp = { source: '^$', options: '' }
     assert_equal expected_regexp, test_regexp.as_json
   end
 
   def test_regexp_to_json
-    assert_equal "/^$/", /\A\Z/.to_json(nil)
+    expected_regexp = { source: '^$', options: '' }
+    assert_equal expected_regexp, /\A\Z/.to_json(nil)
+  end
+
+  def test_regexp_in_hash_to_json
+    expected_regexp = { hello: { source: 'world', options: 'i' } }
+    hash = { hello: /world/i }
+    assert_equal expected_regexp.to_json, hash.to_json
   end
 
   def test_regexp_encode_json
@@ -28,15 +30,18 @@ class CoreExtTest < Test::Unit::TestCase
   end
 
   def test_regexp_remove_comment
-    assert_equal "//", /(?# comment)/.to_json(nil)
+    expected_regexp = { source: '', options: '' }
+    assert_equal expected_regexp, /(?# comment)/.to_json(nil)
   end
 
   def test_regexp_remove_group_options
-    assert_equal "/(something)/", /(?-mix:something)/.to_json(nil)
+    expected_regexp = { source: '(something)', options: '' }
+    assert_equal expected_regexp, /(?-mix:something)/.to_json(nil)
   end
 
-  def test_regexp_as_jason_with_options
-    assert_equal //i, //i.as_json
+  def test_regexp_as_json_with_options
+    expected_regexp = { source: '', options: 'i' }
+    assert_equal expected_regexp, //i.as_json
   end
 
   def test_range_as_json
@@ -54,16 +59,17 @@ class CoreExtTest < Test::Unit::TestCase
   def test_multiline_regexp_as_json
     test_regexp = /
     /
-    expected_regexp = //
+    expected_regexp = { source: '', options: '' }
     assert_equal expected_regexp, test_regexp.as_json
   end
 
   def test_regexp_modifiers_as_json
     # JS allows /i and /m modifiers, all other lead to error
-    assert_equal(//i, //i.as_json)
-    assert_equal(//m, //m.as_json)
-    assert_equal(//im, //im.as_json)
-    assert_equal(//, //x.as_json)
-    assert_equal(//i, //ix.as_json)
+    assert_equal(({ source: '', options: 'i' }), //i.as_json)
+    assert_equal(({ source: '', options: 'm' }), //m.as_json)
+    assert_equal(({ source: '', options: 'im' }), //im.as_json)
+    assert_equal(({ source: '', options: '' }), //x.as_json)
+    assert_equal(({ source: '', options: 'i' }), //ix.as_json)
   end
+
 end
