@@ -17,8 +17,9 @@ class InstallGeneratorTest < Rails::Generators::TestCase
 
   test 'Assert all files are properly created when asset pipeline present and disabled' do
     stub_configuration
-    configuration = {:enabled => false}
+    configuration = {}
     configuration.stubs(:prefix).returns('/assets')
+    ClientSideValidations::Generators::CopyAssetsGenerator.stubs(:asset_pipeline_enabled?).returns false
     Rails.configuration.stubs(:assets).returns(configuration)
     run_generator
     assert_file 'config/initializers/client_side_validations.rb'
@@ -27,8 +28,9 @@ class InstallGeneratorTest < Rails::Generators::TestCase
 
   test 'Assert all files are properly created when asset pipeline present and enabled' do
     stub_configuration
-    configuration = {:enabled => true}
+    configuration = {}
     configuration.stubs(:prefix).returns('/assets')
+    ClientSideValidations::Generators::CopyAssetsGenerator.stubs(:asset_pipeline_enabled?).returns true
     Rails.configuration.stubs(:assets).returns(configuration)
     run_generator
     assert_file    'config/initializers/client_side_validations.rb'
@@ -53,8 +55,9 @@ class CopyAssetsGeneratorTest < Rails::Generators::TestCase
 
   test 'Assert file is properly created when asset pipeline present and disabled' do
     stub_configuration
-    configuration = {:enabled => false}
+    configuration = {}
     configuration.stubs(:prefix).returns('/assets')
+    ClientSideValidations::Generators::CopyAssetsGenerator.stubs(:asset_pipeline_enabled?).returns false
     Rails.configuration.stubs(:assets).returns(configuration)
     run_generator
     assert_file 'public/javascripts/rails.validations.js'
@@ -62,8 +65,9 @@ class CopyAssetsGeneratorTest < Rails::Generators::TestCase
 
   test 'Assert file is properly created when asset pipeline present and enabled' do
     stub_configuration
-    configuration = {:enabled => true}
+    configuration = {}
     configuration.stubs(:prefix).returns('/assets')
+    ClientSideValidations::Generators::CopyAssetsGenerator.stubs(:asset_pipeline_enabled?).returns true
     Rails.configuration.stubs(:assets).returns(configuration)
     run_generator
     assert_file 'app/assets/javascripts/rails.validations.js'
@@ -73,4 +77,3 @@ class CopyAssetsGeneratorTest < Rails::Generators::TestCase
     Rails.stubs(:configuration).returns(mock('Configuration'))
   end
 end
-
